refactor: extract expandable text extension
