+++ app/src/main/java/org/ole/planet/myplanet/utilities/TextViewExtensions.kt
@@ -1,0 +1,67 @@
+package org.ole.planet.myplanet.utilities
+
+import android.text.SpannableStringBuilder
+import android.text.Spanned
+import android.text.TextUtils
+import android.text.method.LinkMovementMethod
+import android.text.style.ClickableSpan
+import android.view.View
+import android.widget.TextView
+import org.ole.planet.myplanet.R
+
+fun TextView.makeExpandable(
+    fullText: CharSequence,
+    collapsedMaxLines: Int = 6,
+    expandLabel: String = context.getString(R.string.show_more),
+    collapseLabel: String = context.getString(R.string.show_less)
+) {
+    var isExpanded = false
+
+    fun refresh() {
+        text = fullText
+        if (!isExpanded) {
+            maxLines = collapsedMaxLines
+            ellipsize = TextUtils.TruncateAt.END
+
+            post {
+                if (lineCount > collapsedMaxLines) {
+                    val lastChar = layout.getLineEnd(collapsedMaxLines - 2)
+
+                    val visiblePortion = SpannableStringBuilder(fullText.subSequence(0, lastChar))
+
+                    visiblePortion.append("… ").also { sb ->
+                        val start = sb.length
+                        sb.append(expandLabel)
+                        sb.setSpan(object : ClickableSpan() {
+                            override fun onClick(widget: View) {
+                                isExpanded = true
+                                refresh()
+                            }
+                        }, start, sb.length, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE)
+                    }
+
+                    text = visiblePortion
+                    movementMethod = LinkMovementMethod.getInstance()
+                }
+            }
+        } else {
+            val expanded = SpannableStringBuilder(fullText).apply {
+                append(" ")
+                val start = length
+                append(collapseLabel)
+                setSpan(object : ClickableSpan() {
+                    override fun onClick(widget: View) {
+                        isExpanded = false
+                        refresh()
+                    }
+                }, start, length, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE)
+            }
+            text = expanded
+            movementMethod = LinkMovementMethod.getInstance()
+            maxLines = Int.MAX_VALUE
+            ellipsize = null
+        }
+    }
+
+    refresh()
+}
